Rename UserMsg component to ActionAlert and update import in AppFooter
